+++ rector.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 use Rector\Core\Configuration\Option;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Rector\Set\ValueObject\SetList;
-// use Rector\Downgrade\Rector\Property\DowngradeTypedPropertyRector;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     // get parameters
@@ -20,9 +19,9 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     // paths to refactor; solid alternative to CLI arguments
     $parameters->set(Option::PATHS, [
         __DIR__ . '/src',
-        __DIR__ . '/vendor/getpop',
-        __DIR__ . '/vendor/pop-schema',
-        __DIR__ . '/vendor/graphql-by-pop',
+        __DIR__ . '/vendor/getpop/*/src',
+        __DIR__ . '/vendor/pop-schema/*/src',
+        __DIR__ . '/vendor/graphql-by-pop/*/src',
     ]);
 
     // is there a file you need to skip?
@@ -38,13 +37,5 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     ]);
 
     // is your PHP version different from the one your refactor to? [default: your PHP version]
-    $parameters->set(Option::PHP_VERSION_FEATURES, '7.2');
-
-    // Commented: Not adding the docBlocks makes PHPStan fail on level 6
-    // // Don't output the docBlocks when removing typed properties
-    // $services = $containerConfigurator->services();
-    // $services->set(DowngradeTypedPropertyRector::class)
-    //     ->call('configure', [[
-    //         DowngradeTypedPropertyRector::ADD_DOC_BLOCK => false,
-    //     ]]);
+    $parameters->set(Option::PHP_VERSION_FEATURES, '7.1');
 };